requirements.txt
azure-common==1.1.28
azure-core==1.29.4
azure-mgmt-core==1.4.0
azure-mgmt-notificationhubs==8.0.0
azure-servicebus==7.11.2
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
isodate==0.6.1
msrest==0.7.1
oauthlib==3.2.2
plyer==2.1.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.8.0
urllib3==2.0.5
